Add support for skipping conversion to CIEXYZ (not complete)
Some tools are colour space agnostic or even encoding
agnostic, by skipping conversion to CIEXYZ when these
tools are used, the rendering time can be significantly
reduced. The video can also be split horizontally and
vertically, and latted merged back, so it is not necessary
to convert the entire video if only parts of it actually
need it.

Because some tools are less agnostic than other tools,
partial conversion to CIEXYZ is also added.

blind-convert must be updated, and all tools most be test

Signed-off-by: Mattias Andrée <maandree@kth.se>