Fix colour distortion
This matrices are not convert to or from the proper YUV colour space both
rather the ones used in videos. srgb_to_yuv was created by converting the
image [sRGB(100%, 0, 0); sRGB(0, 100%, 0); sRGB(0, 0, 100%)] to YUV with
FFmpeg and transposing the matrix. yuv_to_srgb is the inversion of that
matrix, the values in the first column should all be 1 but are not due to
rounding/truncation error.

Signed-off-by: Mattias Andrée <maandree@kth.se>